feat: 사건 정보 위치 변경
- 사건 정보 탭 내부 -> 사건 상세 페이지 상단
- 그에 따른 디렉터리 구조 변경
