+++ src/components/case/common/CaseEmployeeInfoCard.tsx
@@ -1,9 +1,9 @@
 import Card from "@mui/material/Card";
-import CardTitle from "../common/CardTitle.tsx";
+import CardTitle from "../../common/CardTitle.tsx";
 import List from "@mui/material/List";
-import ListProfileItem from "../common/ListProfileItem.tsx";
+import ListProfileItem from "../../common/ListProfileItem.tsx";
 import { useRecoilValue } from "recoil";
-import caseInfoState from "../../../../states/case/info/caseInfoState.tsx";
+import caseInfoState from "../../../states/case/info/caseInfoState.tsx";
 import BadgeIcon from "@mui/icons-material/Badge";
 
 function CaseEmployeeInfoCard() {
