+++ src/components/case/reception/CaseReceptionInfoCard.tsx
@@ -1,11 +1,11 @@
 import Card from "@mui/material/Card";
-import CardTitle from "../common/CardTitle.tsx";
 import CaseReceptionTable from "./table/CaseReceptionTable.tsx";
 import CaseReceptionSearchBox from "./search/CaseReceptionSearchBox.tsx";
 import Box from "@mui/material/Box";
 import { useRecoilValue } from "recoil";
-import caseReceptionAddPopUpOpenState from "../../../../states/case/info/reception/CaseReceptionAddPopUpOpenState.tsx";
 import CaseReceptionAddPopUp from "./table/popup/CaseReceptionAddPopUp.tsx";
+import CardTitle from "../../common/CardTitle.tsx";
+import caseReceptionAddPopUpOpenState from "../../../states/case/info/reception/CaseReceptionAddPopUpOpenState.tsx";
 
 function CaseReceptionInfoCard() {
   const receptionAddPopUpOpen = useRecoilValue(caseReceptionAddPopUpOpenState);
