+++ src/components/case/reception/table/cell/CaseReceptionDeadlineCell.tsx
@@ -1,26 +1,26 @@
-import caseReceptionsState, {
-  CaseReceptionRowType,
-} from "../../../../../../states/case/info/reception/CaseReceptionsState.tsx";
 import * as dayjs from "dayjs";
 import { Dayjs } from "dayjs";
 import { DatePicker } from "@mui/x-date-pickers";
 import { useRecoilState } from "recoil";
 import { produce } from "immer";
-import { toDateValue } from "../../../../../../lib/convert.ts";
 import Box from "@mui/material/Box";
+import caseReceptionsState, {
+  CaseReceptionRowType,
+} from "../../../../../states/case/info/reception/CaseReceptionsState.tsx";
+import { toDateValue } from "../../../../../lib/convert.ts";
 
 type Props = {
   item: CaseReceptionRowType & { editable: boolean };
 };
 
-function CaseReceptionReceivedAtCell({ item }: Props) {
+function CaseReceptionDeadlineCell({ item }: Props) {
   const [receptions, setReceptions] = useRecoilState(caseReceptionsState);
 
   const handleChange = (e: Dayjs | null) => {
     if (item.editable) {
       const newReceptions = produce(receptions, (draft) => {
         const reception = draft.filter((item2) => item2.id === item.id)[0];
-        reception.receivedAt = e?.toDate().toISOString() ?? "";
+        reception.deadline = e?.toDate().toISOString() ?? "";
       });
       setReceptions(newReceptions);
     }
@@ -32,9 +32,11 @@ function CaseReceptionReceivedAtCell({ item }: Props) {
       disabled={!item.editable}
       format="YYYY-MM-DD"
       slotProps={{ textField: { size: "small" } }}
-      defaultValue={dayjs(item.receivedAt)}
+      defaultValue={dayjs(item.deadline)}
       onChange={handleChange}
-      sx={{ width: "100%" }}
+      sx={{
+        width: "100%",
+      }}
     />
   ) : (
     <Box
@@ -45,9 +47,9 @@ function CaseReceptionReceivedAtCell({ item }: Props) {
         height: 40,
       }}
     >
-      {toDateValue(item.receivedAt)}
+      {toDateValue(item.deadline)}
     </Box>
   );
 }
 
-export default CaseReceptionReceivedAtCell;
+export default CaseReceptionDeadlineCell;
