+++ src/components/case/reception/table/cell/CaseReceptionReceivedAtCell.tsx
@@ -1,26 +1,26 @@
-import caseReceptionsState, {
-  CaseReceptionRowType,
-} from "../../../../../../states/case/info/reception/CaseReceptionsState.tsx";
 import * as dayjs from "dayjs";
 import { Dayjs } from "dayjs";
 import { DatePicker } from "@mui/x-date-pickers";
 import { useRecoilState } from "recoil";
 import { produce } from "immer";
-import { toDateValue } from "../../../../../../lib/convert.ts";
 import Box from "@mui/material/Box";
+import caseReceptionsState, {
+  CaseReceptionRowType,
+} from "../../../../../states/case/info/reception/CaseReceptionsState.tsx";
+import { toDateValue } from "../../../../../lib/convert.ts";
 
 type Props = {
   item: CaseReceptionRowType & { editable: boolean };
 };
 
-function CaseReceptionDeadlineCell({ item }: Props) {
+function CaseReceptionReceivedAtCell({ item }: Props) {
   const [receptions, setReceptions] = useRecoilState(caseReceptionsState);
 
   const handleChange = (e: Dayjs | null) => {
     if (item.editable) {
       const newReceptions = produce(receptions, (draft) => {
         const reception = draft.filter((item2) => item2.id === item.id)[0];
-        reception.deadline = e?.toDate().toISOString() ?? "";
+        reception.receivedAt = e?.toDate().toISOString() ?? "";
       });
       setReceptions(newReceptions);
     }
@@ -32,11 +32,9 @@ function CaseReceptionDeadlineCell({ item }: Props) {
       disabled={!item.editable}
       format="YYYY-MM-DD"
       slotProps={{ textField: { size: "small" } }}
-      defaultValue={dayjs(item.deadline)}
+      defaultValue={dayjs(item.receivedAt)}
       onChange={handleChange}
-      sx={{
-        width: "100%",
-      }}
+      sx={{ width: "100%" }}
     />
   ) : (
     <Box
@@ -47,9 +45,9 @@ function CaseReceptionDeadlineCell({ item }: Props) {
         height: 40,
       }}
     >
-      {toDateValue(item.deadline)}
+      {toDateValue(item.receivedAt)}
     </Box>
   );
 }
 
-export default CaseReceptionDeadlineCell;
+export default CaseReceptionReceivedAtCell;
