+++ src/components/case/reception/table/popup/CaseReceptionAddPopUp.tsx
@@ -1,7 +1,4 @@
-import PopUp from "../../../../../common/PopUp.tsx";
-import CloseButton from "../../../../../common/CloseButton.tsx";
 import { useRecoilValue, useSetRecoilState } from "recoil";
-import caseReceptionAddPopUpOpenState from "../../../../../../states/case/info/reception/CaseReceptionAddPopUpOpenState.tsx";
 import FormControl from "@mui/material/FormControl";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
 import MenuItem from "@mui/material/MenuItem";
@@ -11,15 +8,16 @@ import { Button, TextField } from "@mui/material";
 import { DatePicker } from "@mui/x-date-pickers";
 import * as dayjs from "dayjs";
 import { Dayjs } from "dayjs";
-import request, {
-  RequestSuccessHandler,
-} from "../../../../../../lib/request.ts";
-import caseIdState from "../../../../../../states/case/CaseIdState.tsx";
+import caseIdState from "../../../../../states/case/CaseIdState.tsx";
+import { caseReceptionSearchUrlState } from "../../../../../states/case/info/reception/CaseReceptionSearchState.tsx";
+import caseReceptionAddPopUpOpenState from "../../../../../states/case/info/reception/CaseReceptionAddPopUpOpenState.tsx";
 import caseReceptionsState, {
   CaseReceptionRowType,
-} from "../../../../../../states/case/info/reception/CaseReceptionsState.tsx";
-import { caseReceptionSearchUrlState } from "../../../../../../states/case/info/reception/CaseReceptionSearchState.tsx";
-import caseReceptionSizeState from "../../../../../../states/case/info/reception/CaseReceptionSizeState.tsx";
+} from "../../../../../states/case/info/reception/CaseReceptionsState.tsx";
+import caseReceptionSizeState from "../../../../../states/case/info/reception/CaseReceptionSizeState.tsx";
+import request, { RequestSuccessHandler } from "../../../../../lib/request.ts";
+import PopUp from "../../../../common/PopUp.tsx";
+import CloseButton from "../../../../common/CloseButton.tsx";
 
 function CaseReceptionAddPopUp() {
   const [status, setStatus] = useState("incomplete");
